requirements.txt
asgiref==3.8.1
chardet==5.2.0
colorama==0.4.6
Django==5.1.4
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
pluggy==1.5.0
pypdf==5.1.0
pytest==8.3.4
pytest-django==4.9.0
pytest-mock==3.14.0
reportlab==4.2.5
sqlparse==0.5.3
tzdata==2024.2
